+++ src/components/AboutUsSection/FollowUs/styled.js
@@ -4,7 +4,7 @@ export const List = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
-  margin-top: -10px;
+  margin-bottom: 10px;
 `;
 
 export const Icon = styled.img`
