+++ src/components/Footer/FooterSection/styled.js
@@ -26,7 +26,7 @@ export const FooterContainer = styled.div`
     display: flex;
     flex-direction: row;
     justify-content: space-between;
-    margin-bottom: 50px;
+    margin-bottom: 20px;
   }
 `;
 
@@ -37,10 +37,10 @@ export const LogoContainer = styled.div`
 `;
 
 export const SepnotyLogo = styled.img`
-height: 96px;
-    width: 360px;
-    margin: -40px -25px 0px -22px;
-  
+  height: 96px;
+  width: 360px;
+  margin: -40px -25px 0px -22px;
+
   @media screen and (max-width: 455px) {
     height: 96px;
     width: 220px;
@@ -48,23 +48,20 @@ height: 96px;
 `;
 
 export const SepnotyDescription = styled.p`
-   font-size: 14px;
-    width: 398px;
+  font-size: 14px;
+  width: 398px;
   @media screen and (max-width: 455px) {
     font-size: 14px;
     width: 258px;
-    font-weight:500px;
+    font-weight: 500px;
   }
 `;
 
 export const SepnotyDescription1 = styled.p`
-
   @media screen and (max-width: 455px) {
-    display:none;
+    display: none;
   }
 `;
-
-
 
 export const ServiceContainer = styled.div`
   @media screen and (min-width: 768px) {
@@ -84,10 +81,9 @@ export const ServiceHeading = styled.h1`
 `;
 
 export const ServiceItems = styled.p`
-font-size: 14px;
-    margin-top: 10px;
+  font-size: 14px;
+  margin-top: 10px;
   @media screen and (min-width: 768px) {
-   
   }
 `;
 
@@ -98,14 +94,14 @@ export const ContactUsSection = styled.div`
 `;
 
 export const ContactHeading = styled.h1`
-    font-size: 17px;
-    color: #cac7ee;
-    margin-bottom: 16px;
+  font-size: 17px;
+  color: #cac7ee;
+  margin-bottom: 16px;
   @media screen and (max-width: 455px) {
     margin-bottom: 6px;
-    margin-top:10px;
-    text-align:center;
-    text-align:left;
+    margin-top: 10px;
+    text-align: center;
+    text-align: left;
   }
 `;
 
@@ -113,46 +109,45 @@ export const NewsHeading = styled.h1`
   font-size: 17px;
   color: #b44555;
   @media screen and (max-width: 455px) {
-     margin-top:20px;
-     margin-bottom:5px;
+    margin-top: 20px;
+    margin-bottom: 5px;
   }
 `;
 
 export const NewsDesc = styled.p`
-font-size: 14px;
-    color: #969595;
-    padding: 10px 0px 10px;
-    width: 316px;
+  font-size: 14px;
+  color: #969595;
+  padding: 10px 0px 10px;
+  width: 316px;
   @media screen and (max-width: 455px) {
-     display:none;
+    display: none;
   }
 `;
 
 export const InputHolder = styled.input`
-background-color: #4a4848;
-border-top-left-radius: 2px;
-border-bottom-left-radius: 2px;
-border-width: 0px;
-font-size: 12px;
-padding: 7px;
-width: 200px;
-outline: none;
-z-index: 2;
+  background-color: #4a4848;
+  border-top-left-radius: 2px;
+  border-bottom-left-radius: 2px;
+  border-width: 0px;
+  font-size: 12px;
+  padding: 7px;
+  width: 200px;
+  outline: none;
+  z-index: 2;
 
-& ::placeholder {
-  color: #e2dccc;
-}
+  & ::placeholder {
+    color: #e2dccc;
+  }
   @media screen and (min-width: 768px) {
-   
   }
 `;
 
 export const EmailMsg = styled.p`
-font-size: 12px;
-    color: #969595;
-    padding-top: 10px;
+  font-size: 12px;
+  color: #969595;
+  padding-top: 10px;
   @media screen and (max-width: 455px) {
-     margin-bottom:20px;
+    margin-bottom: 20px;
   }
 `;
 
@@ -163,129 +158,136 @@ export const NewsContainer = styled.div`
 `;
 
 export const Button = styled.button`
-background-color: #f93e53;
-    text-align: center;
-    font-size: 12px;
-    border-width: 0px;
-    padding: 7px 15px 7px 15px;
-    border-top-right-radius: 2px;
-    border-bottom-right-radius: 2px;
-    color: #e2dccc;
-    z-index: 2;
+  background-color: #f93e53;
+  text-align: center;
+  font-size: 12px;
+  border-width: 0px;
+  padding: 7px 15px 7px 15px;
+  border-top-right-radius: 2px;
+  border-bottom-right-radius: 2px;
+  color: #e2dccc;
+  z-index: 2;
   @media screen and (min-width: 768px) {
-    
   }
 `;
 
 export const CopywriteContainer = styled.div`
-display: flex;
-    flex-direction: row;
-    align-items: center;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
 
-    & .copywrite-icon {
-      font-size: 10px;
-    }
+  & .copywrite-icon {
+    font-size: 10px;
+  }
   @media screen and (max-width: 455px) {
-     width:100%;
-     margin-left:-20px;
+    width: 100%;
+    margin-left: -20px;
   }
 `;
 
 export const Copyright = styled.img`
-height: 18px;
-width: 18px;
-margin-right: 5px;
+  height: 18px;
+  width: 18px;
+  margin-right: 5px;
   @media screen and (min-width: 768px) {
-    
   }
 `;
 
 export const CopywriteContent = styled.p`
-font-size: 10px;
-    color: #d9d9d9;
+  font-size: 10px;
+  color: #d9d9d9;
   @media screen and (max-width: 455px) {
     font-size: 8px;
   }
 `;
 
 export const BottomContainer = styled.div`
-display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: 50px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin-top: 50px;
   @media screen and (min-width: 768px) {
-    
   }
 `;
 
 export const SectionContainer = styled.div`
-display: flex;
-    flex-direction: row;
+  display: flex;
+  flex-direction: row;
   @media screen and (max-width: 455px) {
-    width:160px;
+    width: 160px;
   }
 `;
 
 export const Sections = styled.p`
-    font-size: 10px;
-    position: relative;
-    color: #ffffff;
-    border-right: 1px solid #ffffff;
-    margin-right: 10px;
-    padding-right: 10px;
-    height: 10px;
+  font-size: 10px;
+  position: relative;
+  color: #ffffff;
+  border-right: 1px solid #ffffff;
+  margin-right: 10px;
+  padding-right: 10px;
+  height: 10px;
   @media screen and (max-width: 455px) {
     font-size: 6px;
-    width:48px;
+    width: 48px;
     margin-right: 0px;
     padding-right: 0px;
-    text-align:center;
+    text-align: center;
   }
 `;
 
 export const Section = styled.p`
-font-size: 10px;
-    color: #ffffff;
+  font-size: 10px;
+  color: #ffffff;
   @media screen and (min-width: 768px) {
-    
   }
 `;
 
 export const FollowUsContainer = styled.div`
   display: flex;
-    flex-direction: column;
+  flex-direction: column;
   @media screen and (max-width: 455px) {
-     display:none;
+    display: none;
   }
 `;
 
 export const FollowUsContainer1 = styled.div`
-  display: flex;
+  @media screen and (min-width: 456px) {
+    display: none;
+  }
   @media screen and (max-width: 455px) {
-     display:flex;
-     flex-direction: column;
-     
+    display: flex;
+    flex-direction: column;
   }
 `;
 
 export const FollowUsSection = styled.ul`
-    display: flex;
-    flex-direction: column;
-    list-style-type: none;
+  display: flex;
+  flex-direction: column;
+  list-style-type: none;
+  padding: 0px;
+  margin: 0px;
   @media screen and (max-width: 455px) {
-    flex-direction: row;
+    flex-direction: column;
     justify-content: flex-start;
   }
 `;
 
 export const ServiceContact = styled.div`
-@media screen and (max-width: 455px) {
-  display:flex;
+  display: flex;
   flex-direction: row;
-  justify-content: space-between;
-  & > div1 {
-    order:3;
+  @media screen and (max-width: 455px) {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    & > div1 {
+      order: 3;
+    }
   }
-}
+`;
 
+export const ServiceCon = styled.div`
+  margin-right: 80px;
+  @media screen and (max-width: 455px) {
+    margin-right: 0px;
+  }
 `;
